set dtype earlier to make code more efficient

+++ src/sparcscore/ml/datasets.py
@@ -111,17 +111,15 @@ class HDF5SingleCellDataset(Dataset):
         self.return_id = return_id
         self.return_fake_id = return_fake_id
         self.stats()
- 
-        
+      
     def add_hdf_to_index(self, current_label, path, current_index_list):       
         try:
             input_hdf = h5py.File(path, 'r')
 
             if self.index_list is not None:
-                index_handle = np.zeros((len(current_index_list), 2))
+                index_handle = np.zeros((len(current_index_list), 2), dtype=np.int64)
                 for i, ix in enumerate(current_index_list):
                     index_handle[i] = input_hdf.get('single_cell_index')[ix]
-                index_handle = index_handle.astype(np.int64)
             else:
                 index_handle = input_hdf.get('single_cell_index') # to float
